+++ app/src/main/java/io/zoemeow/dutschedule/viewmodel/MainViewModel.kt
@@ -334,7 +334,10 @@ class MainViewModel @Inject constructor(
             // TODO: Remember change year and semester here!
             return@ProcessVariable dutAccountRepository.getSubjectSchedule(
                 accountSession.value.data,
-                SchoolYearItem(year = 22, semester = 1)
+                SchoolYearItem(
+                    year = appSettings.value.currentSchoolYear.year,
+                    semester = appSettings.value.currentSchoolYear.semester
+                )
             )
         },
         onAfterRefresh = { saveSettings() }
@@ -345,7 +348,10 @@ class MainViewModel @Inject constructor(
             // TODO: Remember change year and semester here!
             return@ProcessVariable dutAccountRepository.getSubjectFee(
                 accountSession.value.data,
-                SchoolYearItem(year = 22, semester = 1)
+                SchoolYearItem(
+                    year = appSettings.value.currentSchoolYear.year,
+                    semester = appSettings.value.currentSchoolYear.semester
+                )
             )
         },
         onAfterRefresh = { saveSettings() }